backend/requirements.txt
flask
flask-cors
tensorflow-hub
numpy
pillow